refactor(rust): satisfy connection quality gates
