+++ zeroshot-rust/src/native_v2_cli/tests.rs
@@ -10,6 +10,8 @@ use super::*;
 
 #[path = "tests/attach.rs"]
 mod attach_tests;
+#[path = "tests/connections.rs"]
+mod connection_tests;
 #[path = "tests/environment.rs"]
 mod environment_tests;
 #[path = "tests/lifecycle.rs"]
