+++ zeroshot-rust/src/native_v2_cli/tests/parser.rs
@@ -14,66 +14,6 @@ fn parser_exposes_static_help_and_version_commands() {
 }
 
 #[test]
-fn connection_commands_keep_secret_values_out_of_arguments() {
-    let set = parse_native_v2_args(args(&[
-        "connection",
-        "set",
-        "github",
-        "--target",
-        "prod",
-        "--scope",
-        "org",
-        "--field",
-        "GH_TOKEN",
-    ]))
-    .assert_value();
-    let set = match set {
-        NativeV2CliCommand::ConnectionSet(command) => Some(command),
-        _ => None,
-    }
-    .assert_value_with("connection set command");
-    assert_eq!(set.key.as_str(), "github");
-    assert_eq!(set.route.target.as_deref(), Some("prod"));
-    assert_eq!(set.route.scope, ConnectionScope::Org);
-    let expected_fields = [EnvironmentVariableName::new("GH_TOKEN").assert_value()];
-    assert!(matches!(set.input, ConnectionInput::Prompt(fields) if fields == expected_fields));
-
-    assert!(
-        parse_native_v2_args(args(&[
-            "connection",
-            "set",
-            "github",
-            "--field",
-            "GH_TOKEN",
-            "--json-stdin"
-        ]))
-        .is_err()
-    );
-    assert!(
-        parse_native_v2_args(args(&[
-            "connection",
-            "set",
-            "github",
-            "--field",
-            "GH_TOKEN",
-            "--field",
-            "GH_TOKEN",
-        ]))
-        .is_err()
-    );
-    assert!(
-        parse_native_v2_args(args(&[
-            "connection",
-            "set",
-            "github",
-            "--field",
-            "GH_TOKEN=value",
-        ]))
-        .is_err()
-    );
-}
-
-#[test]
 fn durable_observation_accepts_native_resume_and_execution_filters() {
     let watch = parse_native_v2_args(args(&[
         "watch", "run-7", "--target", "prod", "--after", "cloud:12",
